Make build.rs play nice with OSX + homebrew.
Since QT5 and QT4 supply many of the same libs and QT4 is much more
widely used (according to the Formula's keg_only declaration) homebrew
doesn't link QT5 by default. To work around this we need to nudge
pkg-config and cmake in the right direction for them to find QT5.

Fixes #25